KBV-308 - refactor session request decryption & input validation
