+++ lambdas/session/src/main/java/uk/gov/di/ipv/cri/address/api/service/JWTDecrypter.java
@@ -1,4 +1,4 @@
-package uk.gov.di.ipv.cri.address.library.service;
+package uk.gov.di.ipv.cri.address.api.service;
 
 import com.nimbusds.jose.JOSEException;
 import com.nimbusds.jose.JWEObject;
