+++ lambdas/session/src/test/resources/address-cri-test.pk8
