+++ lambdas/session/src/test/resources/signing_ec.crt.pem
