+++ lambdas/session/src/test/resources/signing_ec.pk8
