+++ lib/src/test/java/uk/gov/di/ipv/cri/address/library/service/AddressSessionServiceTest.java
@@ -4,10 +4,6 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.ObjectReader;
-import com.nimbusds.jose.JOSEException;
-import com.nimbusds.jose.JWSAlgorithm;
-import com.nimbusds.jwt.JWTClaimsSet;
-import com.nimbusds.jwt.SignedJWT;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -19,22 +15,15 @@ import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 import software.amazon.awssdk.enhanced.dynamodb.DynamoDbIndex;
 import software.amazon.awssdk.enhanced.dynamodb.DynamoDbTable;
-import uk.gov.di.ipv.cri.address.library.domain.RawSessionRequest;
 import uk.gov.di.ipv.cri.address.library.domain.SessionRequest;
 import uk.gov.di.ipv.cri.address.library.exception.AddressProcessingException;
-import uk.gov.di.ipv.cri.address.library.exception.ClientConfigurationException;
 import uk.gov.di.ipv.cri.address.library.exception.SessionExpiredException;
 import uk.gov.di.ipv.cri.address.library.exception.SessionNotFoundException;
-import uk.gov.di.ipv.cri.address.library.exception.SessionValidationException;
 import uk.gov.di.ipv.cri.address.library.models.CanonicalAddress;
 import uk.gov.di.ipv.cri.address.library.persistence.DataStore;
 import uk.gov.di.ipv.cri.address.library.persistence.item.AddressSessionItem;
 
-import java.io.IOException;
 import java.net.URI;
-import java.security.cert.Certificate;
-import java.security.cert.CertificateEncodingException;
-import java.text.ParseException;
 import java.time.Clock;
 import java.time.Duration;
 import java.time.Instant;
@@ -42,7 +31,6 @@ import java.time.ZoneId;
 import java.util.*;
 
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.notNullValue;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -50,7 +38,6 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
@@ -61,8 +48,6 @@ class AddressSessionServiceTest {
 
     @Mock private DataStore<AddressSessionItem> mockDataStore;
     @Mock private ConfigurationService mockConfigurationService;
-    @Mock private JWTDecrypter mockJWTDecrypter;
-    @Mock private JWTVerifier mockJwtVerifier;
     @Mock private ObjectMapper mockObjectMapper;
     @Captor private ArgumentCaptor<AddressSessionItem> mockAddressSessionItem;
 
@@ -76,12 +61,7 @@ class AddressSessionServiceTest {
         Clock nowClock = Clock.fixed(fixedInstant, ZoneId.systemDefault());
         addressSessionService =
                 new AddressSessionService(
-                        mockDataStore,
-                        mockConfigurationService,
-                        nowClock,
-                        mockJwtVerifier,
-                        mockJWTDecrypter,
-                        mockObjectMapper);
+                        mockDataStore, mockConfigurationService, nowClock, mockObjectMapper);
     }
 
     @Test
@@ -106,130 +86,6 @@ class AddressSessionServiceTest {
         assertThat(
                 capturedValue.getRedirectUri(),
                 equalTo(URI.create("https://www.example.com/callback")));
-    }
-
-    @Test
-    void shouldThrowValidationExceptionWhenSessionRequestIsInvalid() throws IOException {
-        String requestBody = marshallToJSON(Map.of("not", "a-session-request"));
-        when(mockObjectMapper.readValue(requestBody, RawSessionRequest.class))
-                .thenThrow(Mockito.mock(JsonProcessingException.class));
-        SessionValidationException exception =
-                assertThrows(
-                        SessionValidationException.class,
-                        () -> addressSessionService.validateSessionRequest(requestBody));
-        assertThat(exception.getMessage(), containsString("could not parse request body"));
-        verifyNoInteractions(mockConfigurationService);
-    }
-
-    @Test
-    void shouldThrowValidationExceptionWhenRequestClientIdIsInvalid()
-            throws ParseException, JOSEException, JsonProcessingException {
-        String invalidClientId = "invalid-client-id";
-        String testRequestBody = "test-request-body";
-        String configParameterPath = "/clients/" + invalidClientId + "/jwtAuthentication";
-        SignedJWTBuilder signedJWTBuilder = new SignedJWTBuilder().setClientId(invalidClientId);
-        SignedJWT signedJWT = signedJWTBuilder.build();
-        RawSessionRequest rawSessionRequest = createRawSessionRequest(signedJWT);
-        rawSessionRequest.setClientId(invalidClientId);
-        when(mockObjectMapper.readValue(testRequestBody, RawSessionRequest.class))
-                .thenReturn(rawSessionRequest);
-        when(mockJWTDecrypter.decrypt(rawSessionRequest.getRequestJWT())).thenReturn(signedJWT);
-        when(mockConfigurationService.getParametersForPath(configParameterPath))
-                .thenReturn(Map.of());
-        SessionValidationException exception =
-                assertThrows(
-                        SessionValidationException.class,
-                        () -> addressSessionService.validateSessionRequest(testRequestBody));
-        assertThat(exception.getMessage(), containsString("no configuration for client id"));
-        verify(mockConfigurationService).getParametersForPath(configParameterPath);
-    }
-
-    @Test
-    void shouldThrowValidationExceptionWhenRedirectUriIsInvalid()
-            throws ParseException, JOSEException, JsonProcessingException {
-        SignedJWTBuilder signedJWTBuilder =
-                new SignedJWTBuilder().setRedirectUri("https://www.example.com/not-valid-callback");
-        SignedJWT signedJWT = signedJWTBuilder.build();
-        RawSessionRequest rawSessionRequest = createRawSessionRequest(signedJWTBuilder.build());
-        String requestBody = "test request body";
-        when(mockObjectMapper.readValue(requestBody, RawSessionRequest.class))
-                .thenReturn(rawSessionRequest);
-        when(mockJWTDecrypter.decrypt(rawSessionRequest.getRequestJWT())).thenReturn(signedJWT);
-        initMockConfigurationService(signedJWTBuilder.getCertificate(), false);
-
-        SessionValidationException exception =
-                assertThrows(
-                        SessionValidationException.class,
-                        () -> addressSessionService.validateSessionRequest(requestBody));
-        assertThat(
-                exception.getMessage(),
-                containsString(
-                        "redirect uri https://www.example.com/not-valid-callback does not match configuration uri https://www.example/com/callback"));
-    }
-
-    @Test
-    void shouldThrowValidationExceptionWhenJWTIsInvalid()
-            throws JsonProcessingException, ParseException, JOSEException {
-        RawSessionRequest rawSessionRequest = createRawSessionRequest("not a jwt");
-
-        String requestBody = "test request body";
-        when(mockObjectMapper.readValue(requestBody, RawSessionRequest.class))
-                .thenReturn(rawSessionRequest);
-        when(mockJWTDecrypter.decrypt(rawSessionRequest.getRequestJWT())).thenReturn(null);
-
-        SessionValidationException exception =
-                assertThrows(
-                        SessionValidationException.class,
-                        () -> addressSessionService.validateSessionRequest(requestBody));
-        assertThat(
-                exception.getMessage(),
-                containsString("could not parse request body to signed JWT"));
-    }
-
-    @Test
-    void shouldValidateJWTSignedWithECKey()
-            throws IOException, SessionValidationException, ClientConfigurationException,
-                    java.text.ParseException, JOSEException {
-        String requestBody = "test request body";
-        SignedJWTBuilder signedJWTBuilder =
-                new SignedJWTBuilder()
-                        .setPrivateKeyFile("signing_ec.pk8")
-                        .setCertificateFile("signing_ec.crt.pem")
-                        .setSigningAlgorithm(JWSAlgorithm.ES384);
-        SignedJWT signedJWT = signedJWTBuilder.build();
-        RawSessionRequest rawSessionRequest = createRawSessionRequest(signedJWT);
-        when(mockObjectMapper.readValue(requestBody, RawSessionRequest.class))
-                .thenReturn(rawSessionRequest);
-
-        when(mockJWTDecrypter.decrypt(rawSessionRequest.getRequestJWT())).thenReturn(signedJWT);
-        Map<String, String> configMap = standardSSMConfigMap(signedJWTBuilder.getCertificate());
-        configMap.put("authenticationAlg", "ES384");
-        initMockConfigurationService(configMap, false);
-
-        SessionRequest result = addressSessionService.validateSessionRequest(requestBody);
-
-        makeSessionRequestFieldValueAssertions(
-                result, rawSessionRequest, signedJWT.getJWTClaimsSet());
-    }
-
-    private void makeSessionRequestFieldValueAssertions(
-            SessionRequest sessionRequest,
-            RawSessionRequest rawSessionRequest,
-            JWTClaimsSet jwtClaims)
-            throws java.text.ParseException {
-        assertThat(sessionRequest.getAudience(), equalTo(jwtClaims.getAudience().get(0)));
-        assertThat(sessionRequest.getIssuer(), equalTo(jwtClaims.getIssuer()));
-        assertThat(sessionRequest.getSubject(), equalTo(jwtClaims.getSubject()));
-
-        assertThat(sessionRequest.getState(), equalTo(jwtClaims.getStringClaim("state")));
-        assertThat(sessionRequest.getClientId(), equalTo(rawSessionRequest.getClientId()));
-        assertThat(sessionRequest.getClientId(), equalTo(jwtClaims.getStringClaim("client_id")));
-        assertThat(
-                sessionRequest.getRedirectUri(),
-                equalTo(URI.create(jwtClaims.getStringClaim("redirect_uri"))));
-        assertThat(
-                sessionRequest.getResponseType(),
-                equalTo(jwtClaims.getStringClaim("response_type")));
     }
 
     @Test
@@ -400,59 +256,5 @@ class AddressSessionServiceTest {
                 () ->
                         addressSessionService.saveAddresses(
                                 String.valueOf(UUID.randomUUID()), addresses));
-    }
-
-    private Map<String, String> standardSSMConfigMap(Certificate certificate) {
-        try {
-            HashMap<String, String> map = new HashMap<>();
-            map.put("redirectUri", "https://www.example/com/callback");
-            map.put("authenticationAlg", "RS256");
-            map.put("issuer", "ipv-core");
-            map.put(
-                    "publicCertificateToVerify",
-                    Base64.getEncoder().encodeToString(certificate.getEncoded()));
-            return map;
-        } catch (CertificateEncodingException e) {
-            throw new IllegalStateException(e);
-        }
-    }
-
-    private String marshallToJSON(Object sessionRequest) throws IOException {
-        ObjectMapper objectMapper = new ObjectMapper();
-        return objectMapper.writeValueAsString(sessionRequest);
-    }
-
-    private RawSessionRequest createRawSessionRequest(SignedJWT signedJWT) {
-        return createRawSessionRequest(signedJWT.serialize());
-    }
-
-    private RawSessionRequest createRawSessionRequest(String serialisedJWT) {
-        RawSessionRequest rawSessionRequest = new RawSessionRequest();
-        rawSessionRequest.setClientId("ipv-core");
-        rawSessionRequest.setRequestJWT(serialisedJWT);
-        return rawSessionRequest;
-    }
-
-    private void initMockConfigurationService(Certificate certificate) {
-        initMockConfigurationService(standardSSMConfigMap(certificate));
-    }
-
-    private void initMockConfigurationService(
-            Certificate certificate, boolean stubGetAudienceMethod) {
-        initMockConfigurationService(standardSSMConfigMap(certificate), stubGetAudienceMethod);
-    }
-
-    private void initMockConfigurationService(Map<String, String> parameters) {
-        initMockConfigurationService(parameters, true);
-    }
-
-    private void initMockConfigurationService(
-            Map<String, String> parameters, boolean stubGetAudienceMethod) {
-        when(mockConfigurationService.getParametersForPath("/clients/ipv-core/jwtAuthentication"))
-                .thenReturn(parameters);
-        if (stubGetAudienceMethod) {
-            when(mockConfigurationService.getAddressCriAudienceIdentifier())
-                    .thenReturn("test-audience");
-        }
     }
 }
